Move NotImported to optional_imports.py and start using it.
