+++ great_expectations/optional_imports.py
@@ -10,8 +10,29 @@ We also consolidate logic for warning based on version number in this module.
 from __future__ import annotations
 
 import warnings
+from typing import Any
 
 from packaging.version import Version
+
+
+class NotImported:
+    def __init__(self, message: str):
+        self.__dict__["gx_error_message"] = message
+
+    def __getattr__(self, attr: str) -> Any:
+        raise ModuleNotFoundError(self.__dict__["gx_error_message"])
+
+    def __setattr__(self, key: str, value: Any) -> None:
+        raise ModuleNotFoundError(self.__dict__["gx_error_message"])
+
+    def __call__(self, *args, **kwargs) -> Any:
+        raise ModuleNotFoundError(self.__dict__["gx_error_message"])
+
+    def __str__(self) -> str:
+        return self.__dict__["gx_error_message"]
+
+    def __bool__(self):
+        return False
 
 
 def sqlalchemy_version_check(version: str | Version) -> None:
@@ -48,3 +69,15 @@ def is_version_greater_or_equal(
         compare_version = Version(compare_version)
 
     return version >= compare_version
+
+
+# GX optional imports
+SQLALCHEMY_NOT_IMPORTED = NotImported(
+    "sqlalchemy is not installed, please 'pip install sqlalchemy'"
+)
+try:
+    import sqlalchemy
+
+    sqlalchemy_version_check(sqlalchemy.__version__)
+except ImportError:
+    sqlalchemy = SQLALCHEMY_NOT_IMPORTED
